Refactoring script to split roles
