+++ constants.py
@@ -1,0 +1,13 @@
+# Current version of the application
+CURRENT_VERSION = "Pre-Release"
+API_LINK = "https://api.github.com/repos/wcosker/LabatonSlipMerge/releases/latest"
+REPO_LINK = "https://github.com/wcosker/LabatonSlipMerge/releases"
+
+# Icon path
+ICON_PATH = "assets/icon.png"
+
+# GUI Colors
+BG_COLOR = "#bfd1fc"
+BUTTON_COLOR = "#062984"
+TEXT_COLOR = "#002C54"
+HEADER_COLOR = "#dceeff"
