+++ main.py
@@ -1,0 +1,22 @@
+"""
+################################################################################
+#                                                                              #
+#                           <Labaton Slip Merge>                               #
+#                                                                              #
+#                  [Combining PDF's based on user input]                       #
+#                                                                              #
+# -----------------------------------------------------------------------------#
+#                                                                              #
+#   Author        : William Cosker                                             #
+#   Created       : 9/4/2024                                                   #
+#   Last Modified : 9/20/2024                                                  #
+#   Version       : 0.1.1                                                      #
+#                                                                              #
+################################################################################
+"""
+
+from gui import main_gui
+
+if __name__ == "__main__":
+    main_gui()
+
